--- a/src/sparse.jl
+++ b/src/sparse.jl
@@ -80,23 +80,21 @@ function solve(solver::SparseValueIterationSolver, mdp::SparseTabularMDP)
     transition_A_S_S2 = transition_matrices(mdp)
     reward_S_A = reward_matrix(mdp)
     qvals_S_A = zeros(nS, nA)
-    maxchanges_T = zeros(solver.max_iterations)
 
     total_time = 0.0
     for i in 1:solver.max_iterations
         iter_time = @elapsed begin
             qvalue!(mdp, transition_A_S_S2, reward_S_A, v_S, qvals_S_A, _mul_cache)
-            maxchanges_T[i] = _value!(v_S, qvals_S_A)
+            δ = _value!(v_S, qvals_S_A)
         end
         total_time += iter_time
         if solver.verbose
-            @info "residual: $(maxchanges_T[i]), time: $(iter_time), total time: $(total_time) " i
+            @info "residual: $(δ), time: $(iter_time), total time: $(total_time) " i
         end
-        maxchanges_T[i] < solver.belres && break
+        δ < solver.belres && break
     end
-    qvalue!(mdp, transition_A_S_S2, reward_S_A, v_S, qvals_S_A)
-    # Rounding to avoid floating point error noise
-    policy_S = dropdims(getindex.(argmax(round.(qvals_S_A, digits=20), dims=2), 2), dims=2)
+    qvalue!(mdp, transition_A_S_S2, reward_S_A, v_S, qvals_S_A, _mul_cache)
+    policy_S = argmax.(eachrow(qvals_S_A))
 
     return if solver.include_Q
         ValueIterationPolicy(mdp, qvals_S_A, v_S, policy_S)